+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - Implement embed=true query parameter for docs for embedding in presentations
+- Slides: box layout on right
+- Documentation: Lazy iframes
 
 ## [0.2.34] - 2025-10-21